pyproject.toml
[tool.poetry]
name = "chat-bot"
version = "0.1.0"
description = ""
authors = ["hermeschen1116 <108386417+hermeschen1116@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
wheel = "^0.43.0"
torch = "^2.2.2"
datasets = "^2.18.0"
peft = "^0.9.0"
trl = "^0.7.11"
transformers = "4.38.2"
accelerate = "^0.28.0"
bitsandbytes = "^0.42.0"
evaluate = "^0.4.1"
torcheval = "^0.0.7"
flash-attn = "^2.5.6"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
jupyter = "^1.0.0"
tensorboard = "^2.16.2"
wandb = "^0.16.5"
huggingface-hub = "^0.21.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
